// src/api/models/subscriptionEntity.ts
/**
 * Generated by orval v6.10.2 🍺
 * Do not edit manually.
 * API
 * API DOCS
 * OpenAPI spec version: 1.0
 */
import type { SubscriptionEntityCount } from './subscriptionEntityCount';
import type { PaymentEntity } from './paymentEntity';
import type { Distributor } from './distributor';

export interface SubscriptionEntity {
  active?: boolean;
  createdAt?: string;
  duration?: number;
  endDate?: string;
  id: string;
  nextSubscriptionId?: string;
  _count?: SubscriptionEntityCount;
  payment?: PaymentEntity;
  paymentFinalized?: boolean;
  paymentId?: string;
  previousSubscriptionId?: string;
  previousSubscription?: SubscriptionEntity;
  nextSubscription?: SubscriptionEntity[];
  price?: number;
  subscriber?: Distributor;
  subscriberId?: string;
}